Authorisation check included for all API calls

+++ api/mods/center.rb
@@ -8,46 +8,56 @@ module Venuebook
 			end
 
 			get do
-				return Center.search(params[:search])
+				if authorize! :read, Center
+					return Center.search(params[:search])
+				end
 			end
 
 			get '/:id' do
-				center = Center.find(id: params[:id])
+				if authorize! :read, Center
+					center = Center.find(id: params[:id])
 
-				{
-					center: center,
-					venues: center.venues,
-					admin: center.admin
-				}
+					return {
+						center: center,
+						venues: center.venues,
+						admin: center.admin
+					}
+				end
 			end
 
 			post do
-				center = Center.create(params[:center])
-				center.update({code: SecureRandom.hex(6)})
+				if authorize! :create, Center
+					center = Center.create(params[:center])
+					center.update({code: SecureRandom.hex(6)})
 
-				admin           = User.new(params[:admin])
-				admin.center_id = center.id
-				admin.role      = 3
-				admin.save
+					admin           = User.new(params[:admin])
+					admin.center_id = center.id
+					admin.role      = 3
+					admin.save
 
-				center
+					center
+				end
 			end
 
 			put "/:id" do
-				center = Center.find(id: params[:id])
-				center.update(params[:center])
+				if authorize! :update, Center
+					center = Center.find(id: params[:id])
+					center.update(params[:center])
 
-				center.admin.update(params[:admin]) if params[:admin]
+					center.admin.update(params[:admin]) if params[:admin]
 
-				{
-					center: center,
-					admin: center.admin
-				}
+					return {
+						center: center,
+						admin: center.admin
+					}
+				end
 			end
 
 			delete "/:id" do
-				center = Center.find(id: params[:id])
-				center.destroy
+				if authorize! :destroy, Center
+					center = Center.find(id: params[:id])
+					center.destroy
+				end
 			end
 
 		end
